Projects/imgcap/requirements.txt
os
numpy
pillow
pickle
streamlit
tensorflow